telescreen and television require vision + Move eye to shared (#30260)
* telescreen and television require vision

* Move Content.Server.Eye to Shared

* fix popups

* Filthy
