+++ Content.Shared/Eye/Blinding/Components/ActivatableUIRequiresVisionComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Shared.Eye.Blinding.Components
+{
+    [RegisterComponent]
+    public sealed partial class ActivatableUIRequiresVisionComponent : Component;
+}
+
